go 1.13 yacc

+++ go/vt/sqlparser/sql.go
@@ -1389,7 +1389,7 @@ const yyErrCode = 2
 const yyInitialStackSize = 16
 
 //line yacctab:1
-var yyExca = [...]int16{
+var yyExca = [...]int{
 	-1, 0,
 	1, 35,
 	611, 35,
@@ -1658,7 +1658,7 @@ const yyPrivate = 57344
 
 const yyLast = 42089
 
-var yyAct = [...]int16{
+var yyAct = [...]int{
 	523, 77, 3381, 3352, 3392, 3333, 2516, 3238, 2373, 3296,
 	3334, 959, 3239, 411, 3259, 1204, 3150, 8, 3149, 7,
 	3277, 3148, 6, 3221, 3354, 456, 3141, 3147, 5, 2638,
@@ -5870,7 +5870,7 @@ var yyAct = [...]int16{
 	0, 162, 0, 0, 0, 0, 0, 0, 161,
 }
 
-var yyPact = [...]int32{
+var yyPact = [...]int{
 	4204, -1000, -433, -1000, -1000, -1000, -1000, -1000, -1000, -1000,
 	-1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000,
 	-1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000, -1000,
@@ -6217,7 +6217,7 @@ var yyPact = [...]int32{
 	467, 5724, -1000,
 }
 
-var yyPgo = [...]int16{
+var yyPgo = [...]int{
 	0, 2494, 2493, 42, 220, 2492, 222, 2489, 2488, 15,
 	2487, 27, 21, 18, 16, 2486, 2151, 2149, 2147, 2139,
 	2135, 2485, 12, 2484, 2482, 2133, 2481, 2479, 2129, 2115,
@@ -6257,7 +6257,7 @@ var yyPgo = [...]int16{
 	2067, 2065, 0, 1270, 140, 63, 2055, 2046, 2042,
 }
 
-var yyR1 = [...]int16{
+var yyR1 = [...]int{
 	0, 361, 361, 1, 1, 1, 1, 1, 1, 1,
 	1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
 	1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
@@ -6473,7 +6473,7 @@ var yyR1 = [...]int16{
 	225, 362, 363,
 }
 
-var yyR2 = [...]int8{
+var yyR2 = [...]int{
 	0, 1, 2, 1, 1, 1, 1, 1, 1, 1,
 	1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
 	1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
@@ -6689,7 +6689,7 @@ var yyR2 = [...]int8{
 	1, 1, 1,
 }
 
-var yyChk = [...]int16{
+var yyChk = [...]int{
 	-1000, -361, -1, -3, -10, -11, -12, -13, -14, -16,
 	-52, -17, -18, -19, -58, -59, -60, -64, -65, -66,
 	-62, -61, -25, -28, -20, -68, -29, -30, -31, -32,
@@ -7036,7 +7036,7 @@ var yyChk = [...]int16{
 	99, 611, 195,
 }
 
-var yyDef = [...]int16{
+var yyDef = [...]int{
 	-2, -2, 1, 3, 4, 5, 6, 7, 8, 9,
 	10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
 	20, 21, 22, 23, 24, 25, 26, 27, 28, 29,
@@ -7383,7 +7383,7 @@ var yyDef = [...]int16{
 	0, -2, 305,
 }
 
-var yyTok1 = [...]int16{
+var yyTok1 = [...]int{
 	1, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
 	3, 3, 3, 3, 3, 3, 3, 3, 3, 3,
@@ -7399,7 +7399,7 @@ var yyTok1 = [...]int16{
 	3, 3, 3, 3, 118, 3, 130,
 }
 
-var yyTok2 = [...]int16{
+var yyTok2 = [...]int{
 	2, 3, 4, 5, 6, 7, 8, 9, 10, 11,
 	12, 13, 14, 15, 16, 17, 18, 19, 20, 21,
 	22, 23, 24, 25, 26, 27, 28, 29, 30, 31,
@@ -7428,7 +7428,7 @@ var yyTok2 = [...]int16{
 	270, 271, 272, 273, 274, 275,
 }
 
-var yyTok3 = [...]uint16{
+var yyTok3 = [...]int{
 	57600, 276, 57601, 277, 57602, 278, 57603, 279, 57604, 280,
 	57605, 281, 57606, 282, 57607, 283, 57608, 284, 57609, 285,
 	57610, 286, 57611, 287, 57612, 288, 57613, 289, 57614, 290,
@@ -7577,9 +7577,9 @@ func yyErrorMessage(state, lookAhead int) string {
 	expected := make([]int, 0, 4)
 
 	// Look for shiftable tokens.
-	base := int(yyPact[state])
+	base := yyPact[state]
 	for tok := TOKSTART; tok-1 < len(yyToknames); tok++ {
-		if n := base + tok; n >= 0 && n < yyLast && int(yyChk[int(yyAct[n])]) == tok {
+		if n := base + tok; n >= 0 && n < yyLast && yyChk[yyAct[n]] == tok {
 			if len(expected) == cap(expected) {
 				return res
 			}
@@ -7589,13 +7589,13 @@ func yyErrorMessage(state, lookAhead int) string {
 
 	if yyDef[state] == -2 {
 		i := 0
-		for yyExca[i] != -1 || int(yyExca[i+1]) != state {
+		for yyExca[i] != -1 || yyExca[i+1] != state {
 			i += 2
 		}
 
 		// Look for tokens that we accept or reduce.
 		for i += 2; yyExca[i] >= 0; i += 2 {
-			tok := int(yyExca[i])
+			tok := yyExca[i]
 			if tok < TOKSTART || yyExca[i+1] == 0 {
 				continue
 			}
@@ -7626,30 +7626,30 @@ func yylex1(lex yyLexer, lval *yySymType) (char, token int) {
 	token = 0
 	char = lex.Lex(lval)
 	if char <= 0 {
-		token = int(yyTok1[0])
+		token = yyTok1[0]
 		goto out
 	}
 	if char < len(yyTok1) {
-		token = int(yyTok1[char])
+		token = yyTok1[char]
 		goto out
 	}
 	if char >= yyPrivate {
 		if char < yyPrivate+len(yyTok2) {
-			token = int(yyTok2[char-yyPrivate])
+			token = yyTok2[char-yyPrivate]
 			goto out
 		}
 	}
 	for i := 0; i < len(yyTok3); i += 2 {
-		token = int(yyTok3[i+0])
+		token = yyTok3[i+0]
 		if token == char {
-			token = int(yyTok3[i+1])
+			token = yyTok3[i+1]
 			goto out
 		}
 	}
 
 out:
 	if token == 0 {
-		token = int(yyTok2[1]) /* unknown char */
+		token = yyTok2[1] /* unknown char */
 	}
 	if yyDebug >= 3 {
 		__yyfmt__.Printf("lex %s(%d)\n", yyTokname(token), uint(char))
@@ -7704,7 +7704,7 @@ yystack:
 	yyS[yyp].yys = yystate
 
 yynewstate:
-	yyn = int(yyPact[yystate])
+	yyn = yyPact[yystate]
 	if yyn <= yyFlag {
 		goto yydefault /* simple state */
 	}
@@ -7715,8 +7715,8 @@ yynewstate:
 	if yyn < 0 || yyn >= yyLast {
 		goto yydefault
 	}
-	yyn = int(yyAct[yyn])
-	if int(yyChk[yyn]) == yytoken { /* valid shift */
+	yyn = yyAct[yyn]
+	if yyChk[yyn] == yytoken { /* valid shift */
 		yyrcvr.char = -1
 		yytoken = -1
 		yyVAL = yyrcvr.lval
@@ -7729,7 +7729,7 @@ yynewstate:
 
 yydefault:
 	/* default state action */
-	yyn = int(yyDef[yystate])
+	yyn = yyDef[yystate]
 	if yyn == -2 {
 		if yyrcvr.char < 0 {
 			yyrcvr.char, yytoken = yylex1(yylex, &yyrcvr.lval)
@@ -7738,18 +7738,18 @@ yydefault:
 		/* look through exception table */
 		xi := 0
 		for {
-			if yyExca[xi+0] == -1 && int(yyExca[xi+1]) == yystate {
+			if yyExca[xi+0] == -1 && yyExca[xi+1] == yystate {
 				break
 			}
 			xi += 2
 		}
 		for xi += 2; ; xi += 2 {
-			yyn = int(yyExca[xi+0])
+			yyn = yyExca[xi+0]
 			if yyn < 0 || yyn == yytoken {
 				break
 			}
 		}
-		yyn = int(yyExca[xi+1])
+		yyn = yyExca[xi+1]
 		if yyn < 0 {
 			goto ret0
 		}
@@ -7771,10 +7771,10 @@ yydefault:
 
 			/* find a state where "error" is a legal shift action */
 			for yyp >= 0 {
-				yyn = int(yyPact[yyS[yyp].yys]) + yyErrCode
+				yyn = yyPact[yyS[yyp].yys] + yyErrCode
 				if yyn >= 0 && yyn < yyLast {
-					yystate = int(yyAct[yyn]) /* simulate a shift of "error" */
-					if int(yyChk[yystate]) == yyErrCode {
+					yystate = yyAct[yyn] /* simulate a shift of "error" */
+					if yyChk[yystate] == yyErrCode {
 						goto yystack
 					}
 				}
@@ -7810,7 +7810,7 @@ yydefault:
 	yypt := yyp
 	_ = yypt // guard against "declared and not used"
 
-	yyp -= int(yyR2[yyn])
+	yyp -= yyR2[yyn]
 	// yyp is now the index of $0. Perform the default action. Iff the
 	// reduced production is ε, $1 is possibly out of range.
 	if yyp+1 >= len(yyS) {
@@ -7821,16 +7821,16 @@ yydefault:
 	yyVAL = yyS[yyp+1]
 
 	/* consult goto table to find next state */
-	yyn = int(yyR1[yyn])
-	yyg := int(yyPgo[yyn])
+	yyn = yyR1[yyn]
+	yyg := yyPgo[yyn]
 	yyj := yyg + yyS[yyp].yys + 1
 
 	if yyj >= yyLast {
-		yystate = int(yyAct[yyg])
+		yystate = yyAct[yyg]
 	} else {
-		yystate = int(yyAct[yyj])
-		if int(yyChk[yystate]) != -yyn {
-			yystate = int(yyAct[yyg])
+		yystate = yyAct[yyj]
+		if yyChk[yystate] != -yyn {
+			yystate = yyAct[yyg]
 		}
 	}
 	// dummy call; replaced with literal code
